requirements.txt
moviepy
pillow
pydub
python-dotenv
